+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Different second driver support (for mobile/desktop safari)
 - Tabs manipulating methods for desktop in `CoreDriver/PlayDriver`
 - Context manipulating methods for mobile in `MobileDriver`
+- [pytest-rerunfailures](https://pypi.org/project/pytest-rerunfailures/#pytest-rerunfailures) support
 - Type annotations for most of code
 - Auto `locator_type` support for `com.android` locator 
 - `element.hover` support on mobiles
@@ -19,12 +20,13 @@
 ### Changed
 - `Driver` becomes `DriverWrapper` for more readability
 - Mixins classes renamed and moved to `dyatel.mixins` folder
-- Selenium `element.wait_element` now using `visibility_of_any_elements_located`
+- Selenium `core_element.wait_element` now using `is_displayed`
 - Selenium exception stacktrace reduced in most cases
 
 ### Fixed
 - Custom `driver_wrapper`/`driver` for child elements
 - Selenium `KeyError` of `driver_wrapper.set_cookie` without `domain` 
+- Driver creation with function scope of pytest
 
 ---